requirements.txt
gradio>=5.0.0
transformers
torch
sentence-transformers
faiss-cpu
spaces
gTTS
librosa
accelerate
bitsandbytes
sentencepiece
protobuf
marko
